birddog/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.21.0
cyndilib>=0.0.5
Pillow>=9.0.0 
ndi-python
visca_over_ip